revert 0ul

+++ src/ipc/broad_phase/spatial_hash.cpp
@@ -73,7 +73,7 @@ void SpatialHash::build(
         num_vertices, Eigen::Array3i::Zero());
     std::vector<Eigen::Array3i> vertexMaxVAI(
         num_vertices, Eigen::Array3i::Zero());
-    tbb::parallel_for(0ul, num_vertices, [&](size_t vi) {
+    tbb::parallel_for(size_t(0), num_vertices, [&](size_t vi) {
         ArrayMax3d v_min = vertices_t0.row(vi).cwiseMin(vertices_t1.row(vi));
         ArrayMax3d v_max = vertices_t0.row(vi).cwiseMax(vertices_t1.row(vi));
         AABB::conservative_inflation(v_min, v_max, inflation_radius);
@@ -88,7 +88,7 @@ void SpatialHash::build(
 
     pointAndEdgeOccupancy.resize(triStartInd);
 
-    tbb::parallel_for(0ul, num_vertices, [&](size_t vi) {
+    tbb::parallel_for(size_t(0), num_vertices, [&](size_t vi) {
         const Eigen::Array3i &mins = vertexMinVAI[vi], &maxs = vertexMaxVAI[vi];
         assert((mins <= maxs).all());
         pointAndEdgeOccupancy[vi].reserve((maxs - mins + 1).prod());
@@ -103,7 +103,7 @@ void SpatialHash::build(
         }
     });
 
-    tbb::parallel_for(0ul, size_t(edges.rows()), [&](size_t ei) {
+    tbb::parallel_for(size_t(0), size_t(edges.rows()), [&](size_t ei) {
         int eiInd = ei + edgeStartInd;
 
         Eigen::Array3i mins =
@@ -696,7 +696,7 @@ void SpatialHash::detect_edge_vertex_candidates(
     tbb::enumerable_thread_specific<std::vector<EdgeVertexCandidate>> storages;
 
     tbb::parallel_for(
-        tbb::blocked_range<size_t>(0ul, vertex_boxes.size()),
+        tbb::blocked_range<size_t>(size_t(0), vertex_boxes.size()),
         [&](const tbb::blocked_range<size_t>& range) {
             auto& local_candidates = storages.local();
 
@@ -728,7 +728,7 @@ void SpatialHash::detect_edge_edge_candidates(
     tbb::enumerable_thread_specific<std::vector<EdgeEdgeCandidate>> storages;
 
     tbb::parallel_for(
-        tbb::blocked_range<size_t>(0ul, edge_boxes.size()),
+        tbb::blocked_range<size_t>(size_t(0), edge_boxes.size()),
         [&](const tbb::blocked_range<size_t>& range) {
             auto& local_candidates = storages.local();
 
@@ -760,7 +760,7 @@ void SpatialHash::detect_face_vertex_candidates(
     tbb::enumerable_thread_specific<std::vector<FaceVertexCandidate>> storages;
 
     tbb::parallel_for(
-        tbb::blocked_range<size_t>(0ul, vertex_boxes.size()),
+        tbb::blocked_range<size_t>(size_t(0), vertex_boxes.size()),
         [&](const tbb::blocked_range<size_t>& range) {
             auto& local_candidates = storages.local();
 
@@ -792,7 +792,7 @@ void SpatialHash::detect_edge_face_candidates(
     tbb::enumerable_thread_specific<std::vector<EdgeFaceCandidate>> storages;
 
     tbb::parallel_for(
-        tbb::blocked_range<size_t>(0ul, edge_boxes.size()),
+        tbb::blocked_range<size_t>(size_t(0), edge_boxes.size()),
         [&](const tbb::blocked_range<size_t>& range) {
             auto& local_candidates = storages.local();
